Move code back out of __init__ After discussing this I think it is better to only keep initialization code and so on in __init__
